# backend/api-docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AASTU Facilities Management System API
  description: |
    REST API for the AASTU Facilities Management System. This API enables students, staff, 
    maintenance teams, and administrators to efficiently report, track, and resolve campus 
    maintenance issues across Addis Ababa Science & Technology University.
    
    ## Features
    - JWT-based authentication with role-based access control
    - Complete report lifecycle management
    - Photo upload support
    - Duplicate detection system
    - Real-time notifications
    - Analytics and reporting
    
    ## User Roles
    - **Reporter**: Submit and track reports
    - **Coordinator**: Review and approve reports for assigned blocks
    - **Fixer**: Handle assigned maintenance tasks (Electrical/Mechanical)
    - **Admin**: System configuration and user management
  version: 1.0.0
  contact:
    name: AASTU FMS API Support
    email: support@aastu.edu.et
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api-fms.aastu.edu.et/v1
    description: Production server
  - url: https://staging-api-fms.aastu.edu.et/v1
    description: Staging server
  - url: http://localhost:3000
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /reports:
    post:
      tags:
        - Reports
      summary: Submit new report
      description: Submit a new maintenance report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportSubmission'
      responses:
        '201':
          description: Report submitted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          ticket_id:
                            type: string
                            example: "AASTU-FIX-20240320-0045"
                          status:
                            type: string
                            example: "submitted"
                          submitted_at:
                            type: string
                            format: date-time
                          coordinator_assigned:
                            type: object
                            properties:
                              name:
                                type: string
                              email:
                                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Duplicate report detected
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      success:
                        type: boolean
                        example: false
                      data:
                        type: object
                        properties:
                          duplicate_ticket_id:
                            type: string
                            example: "AASTU-FIX-20240320-0043"
                          duplicate_status:
                            type: string
                            example: "in_progress"
                      error_code:
                        type: string
                        example: "REPORT_001"
...
  /admin/users:
    post:
      tags:
        - Admin
      summary: Create user
      description: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - role
                - full_name
              properties:
                email:
                  type: string
                  format: email
                  example: "newuser@aastu.edu.et"
                role:
                  type: string
                  enum: [reporter, coordinator, electrical_fixer, mechanical_fixer, admin]
                full_name:
                  type: string
                  example: "New User"
                assigned_blocks:
                  type: array
                  items:
                    type: integer
                    minimum: 1
                    maximum: 100
                  description: "Required for coordinator role"
                location_not_specified:
                  type: boolean
                  description: "For coordinators handling general location reports"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/blocks:
    post:
      tags:
        - Admin
      summary: Create block
      description: Create a new building block
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - block_number
                - name
              properties:
                block_number:
                  type: integer
                  minimum: 1
                  maximum: 100
                  example: 101
                name:
                  type: string
                  example: "New Science Building"
                description:
                  type: string
                  example: "Newly constructed science facility"
                floors:
                  type: integer
                  example: 5
                coordinator_ids:
                  type: array
                  items:
                    type: string
                  example: ["USR-2024-00124"]
      responses:
        '201':
          description: Block created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          block_id:
                            type: integer
                          block_number:
                            type: integer
                          name:
                            type: string
                          description:
                            type: string
                          floors:
                            type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /notifications:
    get:
      tags:
        - Notifications
      summary: Get notifications
      description: Get user notifications
      parameters:
        - name: unread_only
          in: query
          schema:
            type: boolean
            default: false
        - name: type
          in: query
          schema:
            type: string
            enum: [info, warning, alert]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Notifications retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          notifications:
                            type: array
                            items:
                              $ref: '#/components/schemas/Notification'
                          unread_count:
                            type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /analytics:
    get:
      tags:
        - Analytics
      summary: Get system analytics
      description: Get comprehensive system analytics and performance metrics
      parameters:
        - name: block_id
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: period
          in: query
          schema:
            type: string
            enum: [day, week, month, quarter, year]
            default: month
        - name: metric
          in: query
          schema:
            type: string
            enum: [completion_rate, avg_time, rating, duplicate_rate]
      responses:
        '200':
          description: Analytics data retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          block_performance:
                            type: array
                            items:
                              type: object
                              properties:
                                block_id:
                                  type: integer
                                block_name:
                                  type: string
                                metrics:
                                  type: object
                                  properties:
                                    total_reports:
                                      type: integer
                                    completion_rate:
                                      type: number
                                      format: float
                                    avg_completion_time_hours:
                                      type: number
                                      format: float
                                    avg_rating:
                                      type: number
                                      format: float
                                    duplicate_rate:
                                      type: number
                                      format: float
                          priority_distribution:
                            type: object
                            properties:
                              emergency:
                                type: integer
                              high:
                                type: integer
                              medium:
                                type: integer
                              low:
                                type: integer
                          trends:
                            type: object
                            properties:
                              reports_last_7_days:
                                type: array
                                items:
                                  type: integer
                              completion_rate_trend:
                                type: array
                                items:
                                  type: number
                                  format: float
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
        timestamp:
          type: string
          format: date-time
          example: "2024-01-20T10:30:00Z"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        data:
          type: object
          nullable: true
          example: null
        message:
          type: string
          example: "An error occurred"
        error_code:
          type: string
          example: "ERROR_001"
        timestamp:
          type: string
          format: date-time
    User:
      type: object
      properties:
        id:
          type: string
          example: "USR-2024-00123"
        email:
          type: string
          format: email
          example: "user@aastu.edu.et"
        role:
          type: string
          enum: [reporter, coordinator, electrical_fixer, mechanical_fixer, admin]
        full_name:
          type: string
          example: "John Doe"
        permissions:
          type: array
          items:
            type: string
          example: ["report:create", "report:view_own"]
...
    ReportSubmission:
      type: object
      required:
        - category
        - location
        - equipment_description
        - problem_description
        - photos
      properties:
        category:
          type: string
          enum: [electrical, mechanical]
        location:
          type: object
          required:
            - type
          properties:
            type:
              type: string
              enum: [specific, general]
            block_id:
              type: integer
              minimum: 1
              maximum: 100
              description: "Required if type is 'specific'"
            room_number:
              type: string
              example: "201"
            description:
              type: string
              example: "Pathway between Block 12 and Library"
              description: "Required if type is 'general'"
        equipment_description:
          type: string
          example: "Lab computer #5 in row 3"
          minLength: 5
          maxLength: 200
        problem_description:
          type: string
          example: "Computer won't turn on. No power light."
          minLength: 10
          maxLength: 500
        photos:
          type: array
          items:
            type: string
            format: base64
          minItems: 1
          maxItems: 3
...
    Notification:
      type: object
      properties:
        id:
          type: string
          example: "NOTIF-001"
        type:
          type: string
          enum: [info, warning, alert]
        title:
          type: string
          example: "SLA Warning"
        message:
          type: string
          example: "Emergency report has 30 minutes remaining"
        data:
          type: object
          properties:
            ticket_id:
              type: string
            remaining_minutes:
              type: integer
        read:
          type: boolean
        created_at:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            data: null
            message: "Validation failed"
            error_code: "VALID_002"
            timestamp: "2024-01-20T10:30:00Z"
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            data: null
            message: "Authentication required"
            error_code: "AUTH_002"
            timestamp: "2024-01-20T10:30:00Z"
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            data: null
            message: "Insufficient permissions"
            error_code: "AUTH_003"
            timestamp: "2024-01-20T10:30:00Z"
...